+++ packages/s2-core/__tests__/spreadsheet/miss-dimension-values-spec.ts
@@ -202,10 +202,12 @@ describe('Miss Dimension Values Tests', () => {
   });
 
   test('should get correctly empty dimension values and use custom placeholder text', async () => {
-    const placeholder = '*';
+    const cellPlaceholder = '*';
 
     s2.setOptions({
-      placeholder,
+      placeholder: {
+        cell: cellPlaceholder,
+      },
     });
 
     await s2.render(false);
@@ -213,7 +215,7 @@ describe('Miss Dimension Values Tests', () => {
     const emptyDimensionValueNode = s2.facet.getRowNodes()[0].children[0];
 
     expect(emptyDimensionValueNode.belongsCell!.getActualText()).toEqual(
-      placeholder,
+      cellPlaceholder,
     );
   });