+++ packages/s2-core/src/utils/export/copy/pivot-data-cell-copy.ts
@@ -8,7 +8,6 @@ import {
   zip,
 } from 'lodash';
 import {
-  AsyncRenderThreshold,
   CornerNodeType,
   EXTRA_FIELD,
   VALUE_FIELD,
@@ -153,18 +152,20 @@ export class PivotDataCellCopy extends BaseDataCellCopy {
         // 因为每次 requestIdleCallback 执行的时间不一样，所以需要记录下当前执行到的 this.leafRowNodes 和 this.leafColNodes
         const dataMatrixIdleCallback = (deadline: IdleDeadline) => {
           const rowLength: number = this.leafRowNodes.length;
+
           // requestIdleCallback 浏览器空闲时会多次执行, 只有一行数据时执行一次即可, 避免生成重复数据
-          let count =
-            rowLength >= AsyncRenderThreshold
-              ? AsyncRenderThreshold
-              : rowLength;
+          this.initIdleCallbackCount(rowLength);
 
           while (
             (deadline.timeRemaining() > 0 || deadline.didTimeout) &&
             rowIndex <= rowLength - 1 &&
-            count > 0
+            this.idleCallbackCount > 0
           ) {
-            for (let j = rowIndex; j < rowLength && count > 0; j++) {
+            for (
+              let j = rowIndex;
+              j < rowLength && this.idleCallbackCount > 0;
+              j++
+            ) {
               const row: DataItem[] = [];
               const rowNode = this.leafRowNodes[j];
 
@@ -183,15 +184,18 @@ export class PivotDataCellCopy extends BaseDataCellCopy {
 
                 row.push(dataItem);
               }
-              rowIndex = j;
+              rowIndex++;
               matrix.push(row);
-              count--;
+              this.idleCallbackCount--;
             }
           }
 
-          if (rowIndex === rowLength - 1) {
+          if (rowIndex === rowLength) {
             resolve(matrix);
           } else {
+            // 重置 idleCallbackCount，避免下次 requestIdleCallback 时 idleCallbackCount 为 0
+            this.initIdleCallbackCount(rowLength);
+
             requestIdleCallback(dataMatrixIdleCallback);
           }
         };