+++ packages/s2-react/__tests__/spreadsheet/table-sheet-spec.tsx
@@ -128,7 +128,9 @@ function MainLayout({ callback }: Props) {
     seriesNumber: {
       enable: true,
     },
-    placeholder: '',
+    placeholder: {
+      cell: '',
+    },
     style: {
       dataCell: {
         height: 32,